requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
razorpay==1.4.2
requests==2.32.3
reportlab==4.2.2
fpdf2==2.8.1
Werkzeug==3.0.3
setuptools>=65.0.0
gunicorn